Rename feature from rule to rep, standing for rule/environment/property
